# heyMovie/src/main/resources/openapi.yaml
openapi: 3.0.3

info:
  title: API Gestion des Films
  description: API REST permettant la gestion des films (CRUD).
  version: 1.0.0
  contact:
    name: Support API - Houssam BACAR
    email: houssam.bacar@etudiant.univ-brest.fr

servers:
  - url: http://localhost:12101
    description: Serveur local de développement

tags:
  - name: movies
    description: Gestion des films

paths:

  /movies:
    get:
      tags:
        - movies
      summary: Récupérer tous les films
      operationId: getAllMovies
      responses:
        "200":
          description: Liste des films
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovieDto"
        "500":
          description: Erreur serveur

    post:
      tags:
        - movies
      summary: Créer un film
      operationId: createMovie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieDto"
      responses:
        "201":
          description: Film créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"
        "400":
          description: Données invalides
        "500":
          description: Erreur serveur

  /movies/{id}/open-location:
    put:
      tags:
        - movies
      summary: Ouvrir un film à la location
      operationId: openLocation
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du film
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Film ouvert à la location
        "404":
          description: Film non trouvé
        "409":
          description: Film déjà ouvert à la location
        "500":
          description: Erreur serveur

  /movies/{id}/close-location:
    put:
      tags:
        - movies
      summary: Fermer la location d’un film
      operationId: closeLocation
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du film
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Location fermée
        "404":
          description: Film non trouvé
        "409":
          description: Film déjà fermé
        "500":
          description: Erreur serveur

  /movies/{id}:
    get:
      tags:
        - movies
      summary: Récupérer un film par ID
      operationId: getMovie
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du film
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Film trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDto"
        "404":
          description: Film non trouvé
        "500":
          description: Erreur serveur

    delete:
      tags:
        - movies
      summary: Supprimer un film
      operationId: deleteMovie
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant du film
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Film supprimé
        "404":
          description: Film non trouvé
        "500":
          description: Erreur serveur

components:

  schemas:

    MovieDto:
      type: object
      description: Représentation d’un film
      properties:

        id:
          type: integer
          format: int64
          description: Identifiant unique du film
          example: 1

        titre:
          type: string
          description: Titre du film
          example: Matrix

        annee_realisation:
          type: integer
          description: Année de réalisation
          example: 2021

        prix_location:
          type: number
          format: double
          description: Prix de location
          example: 4.99

        age_minimum:
          type: integer
          description: Age minimum requis
          example: 12

        ouvert_location:
          type: boolean
          description: Indique si le film est disponible à la location
          example: true

        genres:
          type: array
          description: Liste des identifiants de genres
          items:
            type: integer
            format: int64
          example: [1,2]
        poster:
          type: string
          format: byte
          description: Poster du film encodé en base64

        artists:
          type: array
          description: Les artists

      required:
        - titre
        - annee_realisation
        - prix_location
        - age_minimum